Выполнено выделение параметров и методов информации о заданиях в отдельную VM
